Fixing issues in the Import and Export packages
Renamed import and export packages so as to use all lowercase
characters. Have also replaced instances of the utils.InterfaceToMap()
function with explicit type assertion because the previous function was
no longer needed (and is simply too long).

Removed the InterfaceToMap() function in the utils package.
Removed package name from the list of imported packages, commented lines
generated by the Cobra package and renamed the importpkg and exportpkg
folders in git.
Have also altered the `request.FullRequest` method so that it now takes
on the desired path as an argument instead of the complete URL as this
can be derived from the Domain and Port in the request struct.