Removed Tykcommon and moved bundle_test
Have removed references to the Tykcommon package in order fix issue #4.
Have also moved the bundle_test into the `bundle` package and folder to
avoid confusion/clutter (the test is still failing however). Have had to
modify the Travis file accordingly and have added a @mvdan's ci-test.sh
script to the app so as to handle to the tests for each of the packages.