Refactored the underlying packages
Have reorganised the packages that the import and export commands depend
on in an attempt to make these easier to test and change in the future.
This has meant moving the parts of the `utils` package that dealt with
HTTP requests into the `request` package and moving some of the more
generic code from the `importpkg` and `exportpkg` else where so that
they can be shared.

Have also attempted to improve the test coverage.