Changed the `Find` function, added `Edit`&`Delete`
Should resolve issue #7
Changed the `Find` function to be a public function instead of a method
that returns an APIDef object instead of an interface. Have changed the
`Validate` function in accordance.
Have added and `Edit` command and an `Edit` method to the Record
interface so as to allow users to update records in the database.
Updated existing api usage text to reflect changes